--- a/.gitignore
+++ b/.gitignore
@@ -186,3 +186,4 @@ input_images/
 safety-cache/
 falcon-cache/
 output/
+weights-cache/