--- a/cog.yaml
+++ b/cog.yaml
@@ -38,7 +38,7 @@ build:
     - "lpips==0.1.4"
     - "optimum-quanto==0.2.4"
     - "sentencepiece==0.2.0"
-    - "peft==0.12.0"
+    - "peft==0.13.0"
 
     # llava
     - "deepspeed==0.9.5"