Profiled lora loading in prod to see what was taking so long - it looks like our old nemesis `kaiming_uniform`, and other such functions that are used to randomly initialize empty tensors.

conveniently, someone [just pushed functionality to peft](https://github.com/huggingface/peft/pull/1961) to disable this. So we don't have to get too wild to turn it off; we just need to patch `load_lora_into_transformer` to take advantage of that functionality until it's integrated into diffusers.

That's what this PR does. With this I saw lora load times (after download) drop from about [10 seconds](https://replicate.com/p/684y0srg3nrm00cj67sr7bg7am) to about [1.2 seconds](https://replicate.com/p/0kq84cyp9srm60cj68ast2sgyc). Tested locally w/dev and schnell.
<!-- ELLIPSIS_HIDDEN -->


----

> [!IMPORTANT]
> Patch `load_lora_into_transformer` to use `low_cpu_mem_usage=True`, significantly reducing LoRA load times, and update dependencies accordingly.
> 
>   - **Behavior**:
>     - Patch `load_lora_into_transformer` in `lora_loading_patch.py` to use `low_cpu_mem_usage=True`, reducing LoRA load times from ~10s to ~1.2s.
>     - Update `predict.py` to use the patched `load_lora_into_transformer` for `FluxPipeline`, `FluxImg2ImgPipeline`, and `FluxInpaintPipeline`.
>   - **Dependencies**:
>     - Upgrade `peft` version from `0.12.0` to `0.13.0` in `cog.yaml`.
>   - **Misc**:
>     - Add `/weights-cache` to `.dockerignore`.
> 
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=replicate%2Fflux-fine-tuner&utm_source=github&utm_medium=referral)<sup> for f53e34cd174091777d9bc4900cc1d8401b359c5b. It will automatically update as commits are pushed.</sup>


<!-- ELLIPSIS_HIDDEN -->