@andreasjansson 2024-09-27
Did a test-only cog-safe-push, it passed https://github.com/replicate/flux-fine-tuner/actions/runs/11072824989